Move SiteTableViewHeader into it’s own file and add model
