+++ Client/Frontend/Widgets/SiteTableViewController.swift
@@ -8,103 +8,6 @@ import Storage
 struct SiteTableViewControllerUX {
     static let HeaderHeight: CGFloat = 32
     static let RowHeight: CGFloat = 44
-    static let HeaderFont = UIFont.systemFont(ofSize: 12, weight: UIFont.Weight.medium)
-    static let HeaderTextMargin: CGFloat = 16
-    static let TrailingSpace: CGFloat = 12
-    static let ImageWidthHeight: CGFloat = 24
-}
-
-class SiteTableViewHeader: UITableViewHeaderFooterView, NotificationThemeable, ReusableCell {
-
-    var collapsibleState: ExpandButtonState? {
-        willSet(state) {
-            collapsibleImageView.image = state?.image?.tinted(withColor: UIColor.Photon.Blue50)
-        }
-    }
-
-    let titleLabel: UILabel = .build { label in
-        label.font = DynamicFontHelper.defaultHelper.DeviceFontMediumBold
-        label.textColor = UIColor.theme.tableView.headerTextDark
-    }
-
-    private let collapsibleImageView: UIImageView = .build { imageView in
-        imageView.image = ExpandButtonState.down.image?.tinted(withColor: UIColor.Photon.Blue50)
-    }
-
-    private var titleTrailingConstraint: NSLayoutConstraint!
-    private var imageViewLeadingConstraint: NSLayoutConstraint!
-
-    fileprivate let bordersHelper = ThemedHeaderFooterViewBordersHelper()
-
-    override var textLabel: UILabel? {
-        return titleLabel
-    }
-
-    override init(reuseIdentifier: String?) {
-        super.init(reuseIdentifier: reuseIdentifier)
-
-        translatesAutoresizingMaskIntoConstraints = false
-        contentView.addSubviews(titleLabel, collapsibleImageView)
-
-        bordersHelper.initBorders(view: self.contentView)
-        setDefaultBordersValues()
-
-        backgroundView = UIView()
-        imageViewLeadingConstraint = titleLabel.trailingAnchor.constraint(
-            equalTo: collapsibleImageView.leadingAnchor,
-            constant: -SiteTableViewControllerUX.HeaderTextMargin)
-        titleTrailingConstraint = titleLabel.trailingAnchor.constraint(
-            equalTo: contentView.trailingAnchor,
-            constant: -SiteTableViewControllerUX.HeaderTextMargin)
-
-        NSLayoutConstraint.activate([
-            titleLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor,
-                                                constant: SiteTableViewControllerUX.HeaderTextMargin),
-            titleTrailingConstraint,
-            titleLabel.centerYAnchor.constraint(equalTo: contentView.centerYAnchor),
-
-            collapsibleImageView.centerYAnchor.constraint(equalTo: contentView.centerYAnchor),
-            imageViewLeadingConstraint,
-            collapsibleImageView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor,
-                                                           constant: -SiteTableViewControllerUX.TrailingSpace),
-            collapsibleImageView.widthAnchor.constraint(equalToConstant: SiteTableViewControllerUX.ImageWidthHeight),
-            collapsibleImageView.heightAnchor.constraint(equalToConstant: SiteTableViewControllerUX.ImageWidthHeight)
-        ])
-
-        showImage(false)
-        applyTheme()
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    override func prepareForReuse() {
-        super.prepareForReuse()
-        setDefaultBordersValues()
-        applyTheme()
-    }
-
-    func applyTheme() {
-        titleLabel.textColor = UIColor.theme.tableView.headerTextDark
-        backgroundView?.backgroundColor = UIColor.theme.tableView.selectedBackground
-        bordersHelper.applyTheme()
-    }
-
-    func showBorder(for location: ThemedHeaderFooterViewBordersHelper.BorderLocation, _ show: Bool) {
-        bordersHelper.showBorder(for: location, show)
-    }
-
-    func showImage(_ show: Bool) {
-        collapsibleImageView.isHidden = !show
-        titleTrailingConstraint.isActive = !show
-        imageViewLeadingConstraint.isActive = show
-    }
-
-    func setDefaultBordersValues() {
-        bordersHelper.showBorder(for: .top, true)
-        bordersHelper.showBorder(for: .bottom, true)
-    }
 }
 
 /**
